Renamed module to "tableformatter" (all lowercase)
