+++ setup.py
@@ -53,7 +53,7 @@ setup(
     url='https://github.com/python-tableformatter/tableformatter',
     license='MIT',
     platforms=['any'],
-    py_modules = ['TableFormatter'],
+    py_modules = ['tableformatter'],
     keywords='table tabular formatter',
     python_requires='>=3.4',
     extras_require=EXTRAS_REQUIRE,
